+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "clamav-client"
 version = "2.0.0"
 edition = "2021"
-rust-version = "1.56.0"
+rust-version = "1.63.0"
 authors = ["Thorsten Blum <thorsten.blum@toblux.com>"]
 homepage = "https://github.com/toblux/rust-clamav-client"
 repository = "https://github.com/toblux/rust-clamav-client"
@@ -14,20 +14,23 @@ keywords = ["clamav", "clamd", "anitvirus", "async", "tokio"]
 exclude = ["clamd", ".github"]
 
 [dependencies]
-tokio = { version = "1.34.0", default-features = false, features = ["fs", "io-util", "net"], optional = true }
-tokio-stream = { version = "0.1.14", default-features = false, optional = true }
-async-std = { version = "1.12.0", optional = true }
 bytes = { version = "1", optional = true }
+async-net = {version = "2.0.0", optional = true}
+futures-lite = {version = "2.5.0", optional = true}
+async-fs = {version = "2.1.2", optional = true}
+blocking = {version = "1.6.1" }
 
 [dev-dependencies]
-tokio = { version = "1.34.0", features = ["io-std", "macros", "rt"] }
-tokio-util = { version = "0.7.10", features = ["io"] }
-async-std = { version = "1.12.0", features = ["attributes"] }
+async-std = { version = "1.13.0", features = ["attributes"] }
+tokio = { version = "1.42.0", features = ["full"] }
+tokio-util = { version = "0.7.13", features = ["io"] }
 
 [features]
-tokio = ["dep:tokio"]
-tokio-stream = ["tokio", "dep:tokio-stream", "dep:bytes"]
-async-std = ["dep:async-std", "dep:bytes"]
+async = ["dep:bytes", "dep:async-net", "dep:futures-lite", "dep:async-fs", ]
+tokio = ["async"]
+tokio-stream = ["async"]
+async-std = ["async"]
+
 
 [package.metadata.docs.rs]
 features = ["tokio", "tokio-stream", "async-std"]
