+++ src/async_std.rs
@@ -1,141 +1,11 @@
-use async_std::{
-    fs::File,
-    io::{self, ReadExt, WriteExt},
-    net::{TcpStream, ToSocketAddrs},
-    path::Path,
-    stream::{Stream, StreamExt},
-};
+use std::path::Path;
 
-#[cfg(unix)]
-use async_std::os::unix::net::UnixStream;
+use async_fs::File;
+use futures_lite::Stream;
 
-use super::{IoResult, DEFAULT_CHUNK_SIZE, END_OF_STREAM, INSTREAM, PING, PONG, SHUTDOWN, VERSION};
-
-async fn send_command<RW: ReadExt + WriteExt + Unpin>(
-    mut stream: RW,
-    command: &[u8],
-    expected_response_length: Option<usize>,
-) -> IoResult {
-    stream.write_all(command).await?;
-    stream.flush().await?;
-
-    let mut response = match expected_response_length {
-        Some(len) => Vec::with_capacity(len),
-        None => Vec::new(),
-    };
-
-    stream.read_to_end(&mut response).await?;
-    Ok(response)
-}
-
-async fn scan<R: ReadExt + Unpin, RW: ReadExt + WriteExt + Unpin>(
-    mut input: R,
-    chunk_size: Option<usize>,
-    mut stream: RW,
-) -> IoResult {
-    stream.write_all(INSTREAM).await?;
-
-    let chunk_size = chunk_size
-        .unwrap_or(DEFAULT_CHUNK_SIZE)
-        .min(u32::MAX as usize);
-
-    let mut buffer = vec![0; chunk_size];
-
-    loop {
-        let len = input.read(&mut buffer[..]).await?;
-        if len != 0 {
-            stream.write_all(&(len as u32).to_be_bytes()).await?;
-            stream.write_all(&buffer[..len]).await?;
-        } else {
-            stream.write_all(END_OF_STREAM).await?;
-            stream.flush().await?;
-            break;
-        }
-    }
-
-    let mut response = Vec::new();
-    stream.read_to_end(&mut response).await?;
-    Ok(response)
-}
-
-async fn _scan_stream<
-    S: Stream<Item = Result<bytes::Bytes, std::io::Error>>,
-    RW: ReadExt + WriteExt + Unpin,
->(
-    input_stream: S,
-    chunk_size: Option<usize>,
-    mut output_stream: RW,
-) -> IoResult {
-    output_stream.write_all(INSTREAM).await?;
-
-    let chunk_size = chunk_size
-        .unwrap_or(DEFAULT_CHUNK_SIZE)
-        .min(u32::MAX as usize);
-
-    let mut input_stream = std::pin::pin!(input_stream);
-
-    while let Some(bytes) = input_stream.next().await {
-        let bytes = bytes?;
-        let bytes = bytes.as_ref();
-        for chunk in bytes.chunks(chunk_size) {
-            let len = chunk.len();
-            output_stream.write_all(&(len as u32).to_be_bytes()).await?;
-            output_stream.write_all(chunk).await?;
-        }
-    }
-
-    output_stream.write_all(END_OF_STREAM).await?;
-    output_stream.flush().await?;
-
-    let mut response = Vec::new();
-    output_stream.read_to_end(&mut response).await?;
-    Ok(response)
-}
-
-/// Use a TCP connection to communicate with a ClamAV server
-#[derive(Copy, Clone)]
-pub struct Tcp<A: ToSocketAddrs + Send + Sync> {
-    /// The address (host and port) of the ClamAV server
-    pub host_address: A,
-}
-
-/// Use a Unix socket connection to communicate with a ClamAV server
-#[derive(Copy, Clone)]
-#[cfg(unix)]
-pub struct Socket<P: AsRef<Path> + Send + Sync> {
-    /// The socket file path of the ClamAV server
-    pub socket_path: P,
-}
-
-/// The communication protocol to use
-pub trait TransportProtocol {
-    /// Bidirectional stream
-    type Stream: ReadExt + WriteExt + Unpin;
-
-    /// Converts the protocol instance into the corresponding stream
-    fn connect(&self) -> impl std::future::Future<Output = io::Result<Self::Stream>> + Send;
-}
-
-impl<A> TransportProtocol for Tcp<A>
-where
-    A: ToSocketAddrs + Send + Sync,
-    <A as async_std::net::ToSocketAddrs>::Iter: std::marker::Send,
-{
-    type Stream = TcpStream;
-
-    fn connect(&self) -> impl std::future::Future<Output = io::Result<Self::Stream>> + Send {
-        TcpStream::connect(&self.host_address)
-    }
-}
-
-#[cfg(unix)]
-impl<P: AsRef<Path> + Send + Sync> TransportProtocol for Socket<P> {
-    type Stream = UnixStream;
-
-    fn connect(&self) -> impl std::future::Future<Output = io::Result<Self::Stream>> + Send {
-        UnixStream::connect(&self.socket_path)
-    }
-}
+use super::{IoResult, PING, SHUTDOWN, VERSION};
+pub use crate::nonblocking::{scan, send_command, TransportProtocol};
+pub use crate::{Socket, Tcp};
 
 /// Sends a ping request to ClamAV
 ///
@@ -166,7 +36,7 @@ impl<P: AsRef<Path> + Send + Sync> TransportProtocol for Socket<P> {
 ///
 pub async fn ping<T: TransportProtocol>(connection: T) -> IoResult {
     let stream = connection.connect().await?;
-    send_command(stream, PING, Some(PONG.len())).await
+    send_command(stream, PING).await
 }
 
 /// Gets the version number from ClamAV
@@ -196,7 +66,7 @@ pub async fn ping<T: TransportProtocol>(connection: T) -> IoResult {
 ///
 pub async fn get_version<T: TransportProtocol>(connection: T) -> IoResult {
     let stream = connection.connect().await?;
-    send_command(stream, VERSION, None).await
+    send_command(stream, VERSION).await
 }
 
 /// Scans a file for viruses
@@ -262,7 +132,7 @@ pub async fn scan_buffer<T: TransportProtocol>(
 /// An [`IoResult`] containing the server's response as a vector of bytes
 ///
 pub async fn scan_stream<
-    S: Stream<Item = Result<bytes::Bytes, io::Error>>,
+    S: Stream<Item = Result<bytes::Bytes, std::io::Error>>,
     T: TransportProtocol,
 >(
     input_stream: S,
@@ -270,7 +140,7 @@ pub async fn scan_stream<
     chunk_size: Option<usize>,
 ) -> IoResult {
     let output_stream = connection.connect().await?;
-    _scan_stream(input_stream, chunk_size, output_stream).await
+    crate::nonblocking::scan_stream(input_stream, chunk_size, output_stream).await
 }
 
 /// Shuts down a ClamAV server
@@ -289,5 +159,5 @@ pub async fn scan_stream<
 ///
 pub async fn shutdown<T: TransportProtocol>(connection: T) -> IoResult {
     let stream = connection.connect().await?;
-    send_command(stream, SHUTDOWN, None).await
+    send_command(stream, SHUTDOWN).await
 }
